Add support for creating V1/V2 hybrid  torrents, and V2 torrents

+++ src/MonoTorrent.Client/MonoTorrent/EditableTorrent.cs
@@ -91,8 +91,8 @@ namespace MonoTorrent
             set => SetString (InfoDict, NameKey, value);
         }
 
-        public long PieceLength {
-            get => GetLong (InfoDict, PieceLengthKey);
+        public int PieceLength {
+            get => (int) GetLong (InfoDict, PieceLengthKey);
             set => SetLong (InfoDict, PieceLengthKey, value);
         }
